+++ e2e/cases/javascript-api/load-config/fallback.config.mjs
@@ -1,0 +1,7 @@
+export default {
+  source: {
+    define: {
+      CONFIG_FILE_NAME: JSON.stringify('fallback'),
+    },
+  },
+};