+++ e2e/cases/plugin-vue/import-attributes-text/src/Normal.vue
@@ -1,0 +1,3 @@
+<template>
+  <div id="normal-vue">Normal Vue SFC</div>
+</template>