+++ e2e/cases/plugin-vue/import-attributes-text/src/index.js
@@ -1,0 +1,7 @@
+import { createApp } from 'vue';
+import text from './App.vue' with { type: 'text' };
+import Normal from './Normal.vue';
+
+window.vueText = text;
+
+createApp(Normal).mount('#root');