+++ packages/core/src/cli/index.ts
@@ -2,14 +2,13 @@ import { defaultLogger, isDebug } from '../logger';
 import type { LogLevel } from '../types';
 import { setupCommands } from './commands';
 
-const { argv } = process;
-
-function initNodeEnv(command: string | undefined) {
-  if (!process.env.NODE_ENV) {
-    process.env.NODE_ENV =
-      command === 'build' || command === 'preview' ? 'production' : 'development';
-  }
-}
+export type RunCLIOptions = {
+  /**
+   * The command-line arguments to parse, matching the shape of Node.js `process.argv`
+   * @default process.argv
+   */
+  argv?: string[];
+};
 
 function showGreeting() {
   // Ensure consistent spacing before the greeting message.
@@ -24,29 +23,26 @@ function showGreeting() {
 }
 
 // ensure log level is set before any log is printed
-function setupLogLevel() {
+function setupLogLevel(argv: string[]) {
   if (argv.length <= 3) {
     return;
   }
 
   const logLevelIndex = argv.findIndex((item) => item === '--log-level' || item === '--logLevel');
   if (logLevelIndex !== -1) {
-    const level = process.argv[logLevelIndex + 1];
+    const level = argv[logLevelIndex + 1];
     if (level && ['warn', 'error', 'silent'].includes(level) && !isDebug()) {
       defaultLogger.level = level as LogLevel;
     }
   }
 }
 
-export function runCLI(): void {
-  const command = argv[2];
-
-  initNodeEnv(command);
-  setupLogLevel();
+export function runCLI({ argv = process.argv }: RunCLIOptions = {}): void {
+  setupLogLevel(argv);
   showGreeting();
 
   try {
-    setupCommands();
+    setupCommands(argv);
   } catch (err) {
     defaultLogger.error('Failed to start Rsbuild CLI.');
     defaultLogger.error(err);