+++ packages/core/src/configChain.ts
@@ -67,14 +67,17 @@ export const CHAIN_ID = {
     JS_MAIN: 'js',
     JS_WORKER: 'js-worker',
     JS_RAW: 'js-raw',
+    JS_TEXT: 'js-text',
     /** CSS oneOf rules */
     CSS_MAIN: 'css',
     CSS_RAW: 'css-raw',
+    CSS_TEXT: 'css-text',
     CSS_URL: 'css-url',
     CSS_INLINE: 'css-inline',
     /** SVG oneOf rules */
     SVG: 'svg',
     SVG_RAW: 'svg-asset-raw',
+    SVG_TEXT: 'svg-asset-text',
     SVG_URL: 'svg-asset-url',
     SVG_ASSET: 'svg-asset',
     SVG_REACT: 'svg-react',