+++ packages/core/src/index.ts
@@ -6,13 +6,16 @@
 import type * as Rspack from '@rspack/core';
 import { rspack } from '@rspack/core';
 
-export { runCLI } from './cli';
+export { type RunCLIOptions, runCLI } from './cli';
 export { createRsbuild } from './createRsbuild';
 export {
   type ConfigParams,
   defineConfig,
   type LoadConfigOptions,
   type LoadConfigResult,
+  type RsbuildConfigAsyncFn,
+  type RsbuildConfigDefinition,
+  type RsbuildConfigSyncFn,
   loadConfig,
 } from './loadConfig';
 // Core methods