+++ packages/plugin-svelte/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rsbuild/plugin-svelte
 
+## 2.0.1 (2026-07-02)
+
+### New features
+
+- feat(plugin-svelte): support text import attributes by @chenjiahan in https://github.com/web-infra-dev/rsbuild/pull/8058
+
 ## 2.0.0 (2026-06-01)
 
 ### Breaking changes